refactor: update CacheStore and error handling in core modules
- Changed the return type of the `fetch` method in `CacheStore` to `ZIO[Scope, Throwable, Bytes>` for better error handling and resource management.
- Refactored `GravitonError` to use an `enum` structure, enhancing clarity and maintainability of error types.
- Updated the `hasher` method in `Hashing` to return `ZIO[Scope, Throwable, Hasher]`, improving resource management.
- Adjusted various methods in `Hashing` and `InMemoryBlockStore` to utilize scoped resources, ensuring proper cleanup and error handling.
- Enhanced `BinaryAttributeKey` and `BinaryAttributes` to support more robust type handling and validation.
- Introduced `InMemoryCacheStore` for improved caching functionality with expiration management.

+++ modules/core/src/test/scala/graviton/ChunkerSpec.scala
@@ -79,7 +79,12 @@ case object ChunkerSpec extends ZIOSpecDefault:
       val stream = ZStream.fromChunk(Chunk.fromArray(bytes))
       val pack   = TokenPack.fromStrings("none", List("zzz"))
       stream
-        .via(ZPipeline.anchoredCdc(pack, avgSize = Limits.MAX_BLOCK_SIZE_IN_BYTES / 2, anchorBonus = 0))
+        .via(
+          ZPipeline.anchoredCdc(
+            pack, 
+            avgSize = Limits.MAX_BLOCK_SIZE_IN_BYTES / 2.toInt, 
+            anchorBonus = 0
+        ))
         .runCollect
         .map { chunks =>
           assertTrue(chunks.forall(_.bytes.length <= Limits.MAX_BLOCK_SIZE_IN_BYTES))
